Fix notice popup
