+++ client/my-sites/subscribers/components/subscribers-header/subscribers-header.tsx
@@ -1,7 +1,6 @@
 import { Gridicon } from '@automattic/components';
-import { HelpCenter, updateLaunchpadSettings } from '@automattic/data-stores';
+import { HelpCenter } from '@automattic/data-stores';
 import { useLocalizeUrl } from '@automattic/i18n-utils';
-import { useQueryClient } from '@tanstack/react-query';
 import { Button } from '@wordpress/components';
 import { useDispatch as useDataStoreDispatch } from '@wordpress/data';
 import { useState } from '@wordpress/element';
@@ -12,7 +11,7 @@ import NavigationHeader from 'calypso/components/navigation-header';
 import isJetpackCloud from 'calypso/lib/jetpack/is-jetpack-cloud';
 import { useSelector } from 'calypso/state';
 import getIsSiteWPCOM from 'calypso/state/selectors/is-site-wpcom';
-import { getSelectedSiteId, getSelectedSiteSlug } from 'calypso/state/ui/selectors';
+import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import { useAddSubscribersCallback, useMigrateSubscribersCallback } from '../../hooks';
 import { AddSubscribersModal } from '../add-subscribers-modal';
 import { MigrateSubscribersModal } from '../migrate-subscribers-modal';
@@ -36,23 +35,12 @@ export const SubscribersHeader = ( {
 	const localizeUrl = useLocalizeUrl();
 	const { setShowSupportDoc } = useDataStoreDispatch( HELP_CENTER_STORE );
 	const siteId = useSelector( getSelectedSiteId ) ?? null;
-	const selectedSiteSlug = useSelector( getSelectedSiteSlug );
 	const isWPCOMSite = useSelector( ( state ) => getIsSiteWPCOM( state, siteId ) );
 	const addSubscribersCallback = useAddSubscribersCallback( siteId );
-	const migrateSubscribersCallback = useMigrateSubscribersCallback();
+	const migrateSubscribersCallback = useMigrateSubscribersCallback( siteId );
 	const [ showSubscriberModal, setShowSubscriberModal ] = useState< 'none' | 'add' | 'migrate' >(
 		'none'
 	);
-
-	const queryClient = useQueryClient();
-	const completeImportSubscribersTask = async () => {
-		if ( selectedSiteSlug ) {
-			await updateLaunchpadSettings( selectedSiteSlug, {
-				checklist_statuses: { import_subscribers: true },
-			} );
-		}
-		queryClient.invalidateQueries( { queryKey: [ 'launchpad' ] } );
-	};
 
 	const openHelpCenter = () => {
 		setShowSupportDoc( localizeUrl( 'https://wordpress.com/support/paid-newsletters/' ) );
@@ -118,10 +106,9 @@ export const SubscribersHeader = ( {
 				<AddSubscribersModal
 					isVisible={ showSubscriberModal === 'add' }
 					onClose={ closeSubscriberModal }
-					addSubscribersCallback={ ( importError ) => {
+					addSubscribersCallback={ () => {
 						closeSubscriberModal();
-						completeImportSubscribersTask();
-						addSubscribersCallback( importError );
+						addSubscribersCallback();
 					} }
 				/>
 			) }
@@ -129,10 +116,9 @@ export const SubscribersHeader = ( {
 				<MigrateSubscribersModal
 					isVisible={ showSubscriberModal === 'migrate' }
 					onClose={ closeSubscriberModal }
-					migrateSubscribersCallback={ ( selectedSourceSiteId, targetSiteId ) => {
+					migrateSubscribersCallback={ ( selectedSourceSiteId ) => {
 						closeSubscriberModal();
-						completeImportSubscribersTask();
-						migrateSubscribersCallback( selectedSourceSiteId, targetSiteId );
+						migrateSubscribersCallback( selectedSourceSiteId );
 					} }
 				/>
 			) }
